build: use common install location
Headers, libraries, and executables get installed under dist/include,
dist/lib, and dist/bin. The intent is to
1. Make shared libraries easier to find. Previously, executables had
to find them in various places, and it was fragile as things moved.
2. Make includes easier to find. Each library is responsible for
installing its includes over in dist/include/<component>, so other
pieces normally only need to -I a single place.
3. Make it so after a build, you can put dist/<target>/bin on your
path and get the latest of everything.
4. Make it so cross-building results in a single place with all the
libraries and executables that need to be copied to the target.

Changes to install everything under dist/ are simpler than changes
to clean things from under dist/. The make rules to delete empty
directories could be macro-ized or something.

Boost for arm is installed in dist/arm-linux (like everything else).
That makes it simpler (for better or for worse) for components
other than fastrtps to find it. It does mean that fastrtps has to
(arbitrarily, now) be built before anything that needs boost.

'armball' is greatly simplified since it no longer has to do the
hunting and gathering to build the target tarball.