build: improved directory pruning
This mostly changes the "clean" rule:
* There is one clean rule in common_rules.mk (new file) that deletes
  directories listed in CLEAN_DIRS, files listed in CLEAN_FILES, then
  deletes empty directories that are either dist or below, or are under
  the current directory (i.e. src/*).
* project_defs.mk and common_defs.mk set up default CLEAN_FILES and
  CLEAN_DIRS
* Some programs extend those settings.

Explicit deletion of files and directories is not silent, but all
empty-dir pruning is silenced.

A clean should return the directory tree to its state before the build
(of the same architecture).

A makefile should now include either common_defs.mk or project_defs.mk,
but not both (project_defs pulls in common_defs). Similar for _rules.

Known issues:
* Boost is still installed every time fast_rtps is built (for
  arm-linux).
* The commkit makefile awkwardly includes project_defs.mk, then
  common_rules.mk; all the other makefiles match.