test/unit_fastrtps: new, with history tests
This is a new fastrtps (no commkit) test that is intended to probe the
corners of Fast-RTPS that we care about. The first part is history
management; this version runs pubs & subs with "shallow" and "deep"
histories (best effort and reliable).

Some of the behavior is hard to explain (e.g. where either pub or sub
has a shallow history, and the other deep), and the test "passes" if
it looks like we can configure it such that it works how we want. That
does not mean there are not "issues".

One tests (reliable, both histories deep) is failing, since I can't
think of why we shouldn't get all messages through.

Sometimes a test fails because the subscriber gets nothing; something
is truly broken. A better test would identify that separately from
getting the right number of messages (zero is special, after all), but
the test just says "wrong number of messages".