test/unit_fastrtps: force more edge cases, adjust expectations
More protocol exercise is stimulated by reducing the socket buffer
size to unreasonably small (which causes drops and resends).

Test expectations are that when qos is reliable and both the pub and sub
histories are larger than the message bursts, we expect to get all
messages, but otherwise, we really don't know how many will get through.
(In practice, this is okay, since reliable connections can/will have
histories that are plenty deep enough.) The test just makes sure that
_some_ get through, and that traffic does not stop altogether.

Added sequenceNumberSet so that when something fails, there's one more
bit of evidence available for debugging.

These changes are what led to commit 418ca03 in 3drobotics/Fast-RTPS.